[ApplicationPage] 리코일 초기화 위치 변경 (#456)
* refactor: 리코일 초기화 함수 분리

* feat: 닫기 버튼 클릭시 초기화 함수 호출
